move getShortcutIntent into TransportrUtils
