+++ src/de/grobox/liberario/fragments/SettingsFragment.java
@@ -20,7 +20,6 @@ package de.grobox.liberario.fragments;
 import android.app.Activity;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.net.Uri;
 import android.os.Bundle;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.app.ActivityOptionsCompat;
@@ -96,7 +95,7 @@ public class SettingsFragment extends PreferenceFragmentCompat implements Shared
 			public boolean onPreferenceClick(Preference preference) {
 				// create launcher shortcut
 				Intent addIntent = new Intent();
-				addIntent.putExtra(Intent.EXTRA_SHORTCUT_INTENT, MainActivity.getShortcutIntent(getContext()));
+				addIntent.putExtra(Intent.EXTRA_SHORTCUT_INTENT, TransportrUtils.getShortcutIntent(getContext()));
 				addIntent.putExtra(Intent.EXTRA_SHORTCUT_NAME, getString(R.string.widget_name_quickhome));
 				addIntent.putExtra(Intent.EXTRA_SHORTCUT_ICON_RESOURCE, Intent.ShortcutIconResource.fromContext(getContext(), R.drawable.ic_quickhome_widget));
 				addIntent.putExtra("duplicate", false);
